modify database

+++ src/Goods.java
@@ -1,14 +1,14 @@
 class goods
 {
     String goodsname;   //商品名称
-    double price;       //商品价格
+    float price;       //商品价格
     int number;         //商品剩余数量
     //默认初始化商品函数
     public goods()
     {
     }
     //初始化商品函数
-    public goods(String goodsname, double price, int number)
+    public goods(String goodsname, float price, int number)
     {
         this.goodsname = goodsname;
         this.price = price;
@@ -31,7 +31,7 @@ class goods
         this.number = number;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(float price) {
         this.price = price;
     }
 }
